+++ casa/casa/OS/HostInfoDarwin.h
@@ -92,13 +92,13 @@ friend class HostInfo;
     int valid;
     int cpus;
 
-    int memory_total;
-    int memory_used;
-    int memory_free;
+    ssize_t memory_total;
+    ssize_t memory_used;
+    ssize_t memory_free;
 
-    int swap_total;
-    int swap_used;
-    int swap_free;
+    ssize_t swap_total;
+    ssize_t swap_used;
+    ssize_t swap_free;
 };
 
 // </group>
@@ -127,7 +127,7 @@ HostMachineInfo::HostMachineInfo( ) : valid(1) {
     if ( ret != KERN_SUCCESS ) {
 	valid = 0;
     } else {
-	memory_total = (int) (basic_info.memory_size / 1024);
+	memory_total = basic_info.memory_size / 1024;
 	cpus = basic_info.avail_cpus;
     }
 }
